- separated the timestamp and password generation into their own methods for better code organization and re-usability. - included input validation to ensure that the required parameters are provided and have valid values.will help prevent runtime errors and  improve on overall reliability. - also added more detailed error handling for the HTTP requests.
- separation of concerns, response is now in a separate file
- fixed some typos
